# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts for the current company
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate account code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}:
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/seed-defaults:
    post:
      operationId: seedDefaultAccounts
      tags: [accounts]
      summary: Seed the default Egyptian chart of accounts (only when empty)
      responses:
        "201":
          description: Seeded accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Accounts already exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Account:
      type: object
      required: [id, code, nameAr, type, isGroup, createdAt]
      properties:
        id:
          type: string
        code:
          type: string
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        type:
          type: string
          enum: [asset, liability, equity, revenue, expense]
        parentId:
          type: ["string", "null"]
        isGroup:
          type: boolean
        createdAt:
          type: string
    AccountInput:
      type: object
      required: [code, nameAr, type]
      properties:
        code:
          type: string
          minLength: 1
        nameAr:
          type: string
          minLength: 1
        nameEn:
          type: ["string", "null"]
        type:
          type: string
          enum: [asset, liability, equity, revenue, expense]
        parentId:
          type: ["string", "null"]
          format: uuid
        isGroup:
          type: boolean
    AccountUpdate:
      type: object
      properties:
        code:
          type: string
          minLength: 1
        nameAr:
          type: string
          minLength: 1
        nameEn:
          type: ["string", "null"]
        type:
          type: string
          enum: [asset, liability, equity, revenue, expense]
        parentId:
          type: ["string", "null"]
          format: uuid
        isGroup:
          type: boolean
...